requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
python-dotenv==1.0.1
drf-yasg==1.21.7
django-cors-headers==4.3.1
pytesseract==0.3.10
pdfplumber==0.11.0
PyPDF2==3.0.1
openai==1.12.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0
setuptools
